requirements.txt
Django==1.5.4
django-registration==1.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
